requirements.txt
numpy>=1.20.0
tqdm>=4.62.0
requests>=2.25.0
python-dotenv>=0.19.0
regex>=2021.8.0